requirements.txt
django
djangorestframework
markdown
django-filter
django-simple-history
coverage
requests